+++ LibOS/shim/include/shim_vma.h
@@ -29,7 +29,7 @@ struct shim_vma_info {
     int prot;  // memory protection flags: PROT_*
     int flags; // MAP_* and VMA_*
     struct shim_handle* file;
-    off_t file_offset;
+    size_t file_offset;
     char comment[VMA_COMMENT_LEN];
 };
 
@@ -80,7 +80,7 @@ int bkeep_mprotect(void* addr, size_t length, int prot, bool is_internal);
  * atomically checks for overlaps and fails if one is found.
  */
 int bkeep_mmap_fixed(void* addr, size_t length, int prot, int flags, struct shim_handle* file,
-                     off_t offset, const char* comment);
+                     size_t offset, const char* comment);
 
 /*
  * Bookkeeping an allocation of memory at any address in the range [`bottom_addr`, `top_addr`).
@@ -89,17 +89,17 @@ int bkeep_mmap_fixed(void* addr, size_t length, int prot, int flags, struct shim
  * Start of bookkept range is returned in `*ret_val_ptr`.
  */
 int bkeep_mmap_any_in_range(void* bottom_addr, void* top_addr, size_t length, int prot, int flags,
-                            struct shim_handle* file, off_t offset, const char* comment,
+                            struct shim_handle* file, size_t offset, const char* comment,
                             void** ret_val_ptr);
 
 /* Shorthand for `bkeep_mmap_any_in_range` with the range
  * [`PAL_CB(user_address.start)`, `PAL_CB(user_address.end)`). */
-int bkeep_mmap_any(size_t length, int prot, int flags, struct shim_handle* file, off_t offset,
+int bkeep_mmap_any(size_t length, int prot, int flags, struct shim_handle* file, size_t offset,
                    const char* comment, void** ret_val_ptr);
 
 /* First tries to bookkeep in the range [`PAL_CB(user_address.start)`, `aslr_addr_top`) and if it
  * fails calls `bkeep_mmap_any`. `aslr_addr_top` is a value randomized on each program run. */
-int bkeep_mmap_any_aslr(size_t length, int prot, int flags, struct shim_handle* file, off_t offset,
+int bkeep_mmap_any_aslr(size_t length, int prot, int flags, struct shim_handle* file, size_t offset,
                         const char* comment, void** ret_val_ptr);
 
 /* Looking up VMA that contains `addr`. If one is found, returns its description in `vma_info`.
